+++ src/main/java/com/vnano/defenders/enchantment/ModEnchantments.java
@@ -5,7 +5,6 @@ import net.minecraft.item.ItemStack;
 
 public final class ModEnchantments {
     public static EnchantmentDefender FOOTWORK;
-    public static EnchantmentDefender FORTIFICATION;
     public static EnchantmentDefender REPRISAL;
     public static EnchantmentDefender FINESSE;
     public static EnchantmentDefender REFLEXES;