+++ src/main/java/com/vnano/defenders/registry/ModContent.java
@@ -39,7 +39,6 @@ public final class ModContent {
     @SubscribeEvent
     public static void registerEnchantments(RegistryEvent.Register<Enchantment> event) {
         ModEnchantments.FOOTWORK = new EnchantmentDefender("footwork", Enchantment.Rarity.UNCOMMON, 3);
-        ModEnchantments.FORTIFICATION = new EnchantmentDefender("fortification", Enchantment.Rarity.UNCOMMON, 3);
         ModEnchantments.REPRISAL = new EnchantmentDefender("reprisal", Enchantment.Rarity.RARE, 3);
         ModEnchantments.FINESSE = new EnchantmentDefender("finesse", Enchantment.Rarity.RARE, 3);
         ModEnchantments.REFLEXES = new EnchantmentDefender("reflexes", Enchantment.Rarity.UNCOMMON, 3);
@@ -47,13 +46,21 @@ public final class ModContent {
         ModEnchantments.SIXTH_SENSE = new EnchantmentDefender("sixth_sense", Enchantment.Rarity.VERY_RARE, 1);
         event.getRegistry().registerAll(
             ModEnchantments.FOOTWORK,
-            ModEnchantments.FORTIFICATION,
             ModEnchantments.REPRISAL,
             ModEnchantments.FINESSE,
             ModEnchantments.REFLEXES,
             ModEnchantments.DEFLECTION,
             ModEnchantments.SIXTH_SENSE
         );
+    }
+
+    @SubscribeEvent
+    public static void ignoreRemovedEnchantments(RegistryEvent.MissingMappings<Enchantment> event) {
+        for (RegistryEvent.MissingMappings.Mapping<Enchantment> mapping : event.getMappings()) {
+            if (new net.minecraft.util.ResourceLocation(DefendersMod.MOD_ID, "fortification").equals(mapping.key)) {
+                mapping.ignore();
+            }
+        }
     }
 
     @SubscribeEvent